python_packages/requirements.txt
minio
loguru
fastapi
uvicorn
PyYAML
python-multipart
httpx
mariadb==1.1.10